refactor(emails): ♻️  move emails to their own package
