+++ apps/web/package.json
@@ -13,6 +13,7 @@
     "@hookform/resolvers": "3.3.4",
     "@mindfulyze/utils": "workspace:*",
     "@mindfulyze/editor": "workspace:*",
+    "@mindfulyze/emails": "workspace:*",
     "@mindfulyze/database": "workspace:*",
     "@next-auth/prisma-adapter": "1.0.7",
     "@paralleldrive/cuid2": "2.2.2",
@@ -26,7 +27,6 @@
     "@radix-ui/react-separator": "1.0.3",
     "@radix-ui/react-slot": "1.0.2",
     "@radix-ui/react-switch": "1.0.3",
-    "@react-email/components": "0.0.14",
     "@supabase/supabase-js": "2.39.3",
     "@vercel/analytics": "1.1.2",
     "@vercel/speed-insights": "1.0.6",
@@ -47,12 +47,10 @@
     "react": "18.2.0",
     "react-day-picker": "8.10.0",
     "react-dom": "18.2.0",
-    "react-email": "2.0.0",
     "react-hook-form": "7.49.3",
     "rehype-autolink-headings": "7.1.0",
     "rehype-slug": "6.0.0",
     "remark-gfm": "4.0.0",
-    "resend": "3.0.0",
     "sonner": "1.3.1",
     "use-debounce": "10.0.0",
     "zod": "3.22.4"
