+++ packages/transactional/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "config/tsconfig/react-library.json",
+  "include": ["."],
+  "exclude": ["dist", "build", "node_modules"],
+  "compilerOptions": {
+    "strict": true,
+    "jsx": "react-jsx"
+  }
+}
