+++ packages/transactional/utils/generateId.ts
@@ -1,0 +1,3 @@
+import { customAlphabet } from 'nanoid'
+
+export const generateId = customAlphabet('0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz', 10) // 10-character random string
